0036 Valid Sudoku
I liked this problem a lot! Row/col verification can be done either
through indexes, or the cleaner way of using board.map row[i] to get
the col, and the row is just board[i].

As for the boxes, using while loops was quite dirty and made me require
using the += 3, when I could have just iterated through an array of
[0, 3, 6] instead. All in all, good simulation problem for working with
matrixes.