0022 Generate Parentheses
This is a backtracking decision-tree type problem. It's combination, so
order doesn't matter, and it's many answers. At each decision in the
tree, you can either decide to add an open paren or closing paren. Then
to ensure validity, all you need to do is count the # of open/close
parentheses. I had a bit of trouble because I did premature optimization,
without realizing it was shortcutting the exploration of paths. I'm not
sure why this is in the Stack section of Leetcode-- it seems more like a
backtracking/DFS type of problem,  honestly.