0853 Car Fleet
This one actually gave me a ton of trouble! I got the majority of the
way there, but lost confidence in my nlogn solution because I had to
sort and spent a lot of time searching for a better solution when it
didn't exist.

After sorting, it took me more time to realize I had to go from the
highest position to the lowest to handle delayed cars. Then I realized
I misunderstood the question (cars don't stop at the finish line), so I
had to remove ceil when calculating rounds. Then cleaning up the code
was a fun matter as well.

I don't think this was really a stack type of problem-- it felt more
like a simulation problem? Lots of reasoning and visualizing the
problem on a line. The most important part was realizing what makes
one set of cars a fleet or not. Aka, only the highest position car
matters, it'll always have an equal or higher rounds timing in a
fleet due to lower speed.