0021 Merge Two Sorted Lists
This is essentially the "merge" part of a mergesort function on two
linked lists. It took a bit of brainpower to think about splicing the
nodes together "in-place" but it happened pretty much naturally. The
use of a dummy node makes this a lot cleaner, just like how using
a dummy node instead of prev makes deletion problems inside of LL
a lot easier as well.