0143 Reorder List
This is an amazing LL problem but I struggled a lot. For my initial
solution, it wasn't too bad to implement but it had O(N) space
complexity. Both had O(N) time complexity.

Then when implementing the O(1) space implementation, I struggled a
whole bunch with edge cases. I honestly could have walked through
examples cleaner. This was honestly just laziness-- I knew that the
connection had to be cut, but I wanted to see if I could do the problem
without bothering, and that was a bad idea.

For the O(1) space implementation, it uses a lot of LL crucials
- Fast/slow pointer (and the trick to stop before middle)
- Reversing a LL
- Merging two linked lists

For the O(N) space implementation, I enjoyed how slightly different
intuitions resulted in the merge happening in different ways.