0138 Copy List With Random Pointer
The O(N) space solution is super easy... but I got led on a wild goose
chase by ChatGPT after forgetting a simple movement step.

The O(1) space solution is tricky and worth doing, interleaving is
super clever and relies on the fact that the random pointer goes
directly to the node! Not too bad to code, honestly.

I think the O(1) space solution is more of a hard than a medium, to be
completely honest. Very clever LL problem-- spent an absurd amount
of time on this.

1. Interleaving
2. Replacing random
3. Un-interleaving
- Original LL must return to original state by the end!