0287 Find the Duplicate Number
This is kind of a LL problem in a way-- think about the values as
pointers to a certain index. So looking for the start of a cycle in a
LL. This problem is kind of impossible without knowing Floyd's cycle
detection algorithm beforehand, but at least the algorithm itself is
rather simple.
1. Fast/slow, find intersection point
2. New slow/slow, find intersection: that's the start of cycle