0146 LRU Cache
A hashmap + LL problem!
For this one, the implementation was pretty tricky and there are a lot
of edge cases to deal with. An important point to know when key vs
value is used and what they're used for. Also, focus on understanding
the problem completely before starting the plan or coding by working
through the example.

This code could have been cleaned up a lot more by utilizing more
helper methods to reduce redundant code.