0023 Merge K Sorted Lists
For the heap solution, remember to advance the actual LL, not a pointer
copy.

The other solution that doesn't utilize a heap is pretty clever and
intuitive-- it's just like merge sort! The merge operation is
linear (number of elements in each list), but you only have to do it
log (number of lists) times.