0875 Koko Eating Bananas
Classic binary search problem, supposedly! It's just a regular binary
search, but on a range of numbers instead. I think the trick here is
to not get trapped thinking of a better solution than
N: (length of the piles array) * log(M: maximum of piles array).

For binary search problems, I find it very helpful to put the target
first when stating conditions, and to put the condition if/else into
words so I don't get confused. Essentially the gist of these problems
is that you want to reduce the range of the search by half at every
iteration.