0153 Find Minimum in Rotated Sorted Array
Binary search problem! Reminds me a lot of, "find target in rotated
sorted array". This problem relies more on understanding which side to
search. E.g, if you're at the middle and the right half is sorted,
the middle is the lowest value of the right half, so you can search the
left. If the entire array is sorted, you search the left. If the
left of the array is sorted (but the entire array isn't sorted), you
sort the right (the minimum is somewhere there).