0567 Permutation in String
Another sliding window problem! This time, it's a fixed sliding window.
This problem is very fun and fair-- after thinking about the brute
force approach, think about how does one reduce work? Like for the
palindromic partitioning problem, you reduce work by returning
early when there's an invalid palindrome. For this problem, you realize
that a permutation is the string in EVERY order, so as long as a window
of s2 has all chars of s1, it's a permutation of it!

I solved it without using the fixed length of the window as a
constraint, but I did it using the fixed size sliding window logic
as well. You just gotta when it's the first fixed size sliding window
problem you've seen, right?