0239 Sliding Window Maximum
Challenging fixed size sliding window problem! The sliding window part
of this problem is actually super easy, the challenging part was
figuring out the montonic decreasing deque.
It's my first time seeing that data structure, so it was rather
challenging. But once I understood, coding it was relatively simple.
Then I hit the edge case of, "how do you know a window has passed an
element?" and that was pretty straightforward to reason about and use
indexes instead of the number itself.

I was thinking a lot about max stacks and monotonic stacks when thinking
about this problem, but I guess I had to think about a monotonic queue!
It seems like when you have a need for ever increasing or decreasing
values to minimize work (e.g, finding a greater or lesser value)
invalidates the need for many previous values, a montonic data structure
is the way to go.