0543 Diameter of Binary Tree
Another binary tree DFS problem! Very similar to binary tree maximum
path sum. The tricky part is dealing with the 1 node case: the diameter
is 0, not 1! Well, I guess it isn't tricky because the natural code
deals with it, but I was panicking that my code wouldn't work, then it
did!