0958 Check Completeness of a Binary Tree
Got this during the 5/30/25 mock interview! I can't believe I took so
long to get this one... I originally randomly thought about a DFS
approach, when reading through the problem-- it clearly requires a
level order traversal aka BFS.

When I coded this out, the interviewer didn't understand my approach.
Instead of using a flag to check if nil has been seen, then returning
false if seen_nil = true and there's another character, I simply did:
if current.nil? and queue[0] is not nil. Which simply means: if this
current node is nil and is another element in queue that's not nil,
return false. If the queue is empty, it'll simply be nil!