0283 Move Zeroes and 781 Rabbits in Forest
Both of these are "walk through the example" type problems.
FOr 283, once you realize all you have to do is swap towards the end
and count the # of continuous zeroes, it's quite easy to implement and
solve.

For 781, walking through both the examples given [1, 1, 2] and
[10, 10, 10] then thinking about what happens during an "edge case", aka
what if there are more than 11 rabbits claiming 10, that's enough to
solve the problem.