0222 Count Complete Tree Nodes
This one is a good DFS problem. I think it should be a medium instead
of an easy for the less than O(N) solution-- it requires a lot of
intuition. E.g, complete binary tree as 2^N - 1. Then if it's not a
perfect complete binary tree, you add current node (1) + left + right,
where this works because a single node is a perfect complete binary
tree and nil is just 0.