0098 Validate Binary Search Tree
This one was a very tricky BST DFS problem! THe intuition to this problem
was that ALL nodes left/right have to be less than/greater. So
essentially, you had to create a "range" of valid values for each node.
E.g, the root could be any value, then the node directly to the left
could only be min to node.val, while the right could only be node.val to
max, and so on.