0230 Kth Smallest Element in a BST
I immediately thought of heaps, but that wasn't the right answer!
This problem is very similar to, "convert sorted array to BST"-- it
makes you realize that by doing in-order traversal on a BST, you get
a sorted array. So you just need to do in-order and return the kth
element visited!

Though I guess for convert sorted array to BST, it's more of the
intuition that selecting the middle value to be the current node then
recursively setting the left/right creates it, is more of the "spark"
that solves the problem than realizations about in-order traversal.