0212 Word Search II
Fun trie and backtracking problem! Implementation wasn't too bad once
it was figured out, but Ruby is too slow and times out on the repeated
characters test, sadly. Translating the code to C++/Java makes it work
immediately, so I guess it's not a time complexity issue.

Learned how to implement a Trie very cleanly and quickly, since I
rewrote this problem so many times.