0040 Combination Sum II
Think about it in terms of a decision tree
Your path either has a 1, or no more 1's
Then for the 1's path, your path either has another 1, or no more 1's.
This is quite easy to implement because of index > start, aka when
start == index, the number can still be used. This means that we get to
use each number the amount of times they appear in candidates, but
without treading down duplicate paths. Clever!

Very similar to Combination Sum, a combination backtracking problem.
Difference is sorting + "skipping" duplicates per recursive level.
E.g, think about it in terms of a decision tree.