0090 Subsets II
This problem is basically the same as combination sum ii. It's more or
less a backtracking combination DFS problem, but with the spin of
having to deal with duplicates. To deal with duplicates effectively,
you need to sort + skip over duplicates on the same recursive level.
This makes it so that if you get [1, 2], and you see another two in the
future, you don't put [1, 2] in again.

This pattern is also present in 3sum as well.