0051 N-Queens
I enjoyed this problem a lot! Made a few minor mistakes, but it was
all quite doable-- took me roughly an hour to solve it. It's a pity
that Ruby is too slow to finish the final 9-case...Anyhow, this is
really a solid backtracking problem!

I think all the non-backtracking solutions are more ideal, though. But
regardless, this was good for learning.