0703 Kth Largest Element in a Stream
Fun heap problem! This kind of relies on the inversion principle as
you're thinking about the problem-- how do I get the Kth largest
quickly? And you soon come to the conclusion that even though you
want the largest, a min heap is better. With a min heap of size K,
the element at the top of the stack is the kth largest. All elements
smaller are popped off, elements larger go into the min heap.