1046 Last Stone Weight
Another heap problem, but with a max heap this time! This problem
is honestly the most straightforward of the heap problems? It's
quite literally: just utilize a heap, grab two largest values... For
the kth largest element, at least you had to keep it within K size
and think about how you'd keep the kth largest at the top (by utilizing
a min heap) when you want the largest.