0973 K Closest Points to Origin
Heap problem! This is a kth largest/smallest pattern type of problem.
It's essentially the same problem as kth largest in stream, but with
the opposite type of heap, the Euclidean distance algo, and storing
"custom" values within the heap. That's what replace/offer in a heap
is for haha...