0621 Task Scheduler
Max heap problem! Keep track of the largest # of labeled tasks per task,
then have a queue to decide when to push available tasks back onto the
heap. Clever use of "time", idle == nothing in heap to use, but the
queue isn't empty. THe max heap + queue is superrr clever.

Remember that shift in Ruby is O(1)!