0815 Bus Routes
I recognize a big flaw in myself-- I find it really hard to think when
I'm flustered or when people are talking and distracting me.

This was a graph problem solved using BFS.

With two ways to tackle the problem
1. Nodes as routes
2. Nodes as stops

With nodes as routes, it fails for me on test #47 on Leetcode because
pre-constructing the route graph is very computationally intensive for
sparse maps.

With nodes as stops, it passes for me. Honestly, nodes as routes is
easier to reason about, but stops isn't that much more complicated.
It's really the same problem as "flight travel destinations" or
whatever, but with slightly more to reason about. Aka, stops on
the same bus don't cost us anything, we only want to see amount of
buses we get on... and managing having many stops to deal with going to
many other stops.