0253 Meeting Rooms II
Interval problem! A spin on the first meeting rooms. I guess all
interval problems involve some sort of sorting, then manipulation
of the start or end time of the interval.

This problem uses a min heap to keep track of room ending times. So
you know when a room is freed, aka you don't need another room. My
initial logic of, "create the new rooms manually, look for the room
with the most recent ending date to place" was sound, just inefficient.
With a heap, it simplifies the problem quite a bit. Something in the
heap == a room being used, then if current time is >= end time, that
means that room has opened and you release all open rooms. Nice!