0057 Insert Interval
Another interval problem. These sort of problems are pretty challenging
for me, I need to work my way through them all. This is an interval
problem with the classic merge intervals algorithm, I think. A big
realization here is when working with intervals, it's much easier to
specify when intervals will NOT overlap, then put the overlap cases
all into an "else". Otherwise, it's a pain to code out and manage
that logic. Remember to use that else!