0056 Merge Intervals
Very similar to 57 insert intervals, but arguably easier since you
need to keep track of less things. Same ol' sort, then utilize the
fact that it's intervals so only need to deal with start time, end
time, or both to manipulate. For this problem, only need to deal with
end times to see when something has to be merged, and only the last
element of output is needed every time... so instead of using prev,
can use the last element of the result/output array.