0435 Non-overlapping intervals
Another fun interval problem! As usual... sort, then some of
manipulation dealing with the properties of ranges. (E.g, figuring out
when it overlaps easily). I was going to give up, looked at the NC video,
saw that it was 12 minutes long, then exited out and solved the problem
on my own rather easily. So it was a matter of giving up in my head,
rather than being unable to solve the problem.

Basically drawing out an example, realize that once it's sorted, all you
need to do is remove the overlapping ranges w/ bigger ends because
it has the potential to overlap with more stuff. That's it.