1851 Minimum Interval to Include Each Query
This was a very difficult interval problem requiring use of a heap,
which reminds me of Meeting Rooms II. The super simple solution is
basically N^2 while this one is nlogn + qlogq, which is quite a bit
better I guess.

The intuition to solve this problem is that you need to get the smallest
size in a quick time, and a heap is perfect for that. Then you choose
the end time as a limiter to decide which items to actually keep
within the heap. So if queries and intervals are both sorted...
-> You remove intervals that are invalid (that are top of the heap)
(Ones behind don't matter, you'll get a valid answer first, and they'll
be popped in the future anyway if necessary).
-> You push all possible intervals into the heap for the first query
-> You take the smallest size to put into result
-> Rinse and repeat-- you don't ever need to readd an interval, only
remove them as they become invalid for future queries. Since you know
that earlier start time == start time is valid for all future queries,
so you'd only need to check the end.