0055 Jump Game
Greedy problem! I did it in two ways

1. Forward greedy
- This solution was much more intuitive and I got it right away, but
I couldn't code it up for the longest time for some reason. I think
I had to focus more on the return false case, and what was the bare
minimum item to actually keep track of.

2. Backwards greedy
- Clever! Moving the goalpost since if you can reach the end from a
previous value, that value can become the new goalpost.