0045 Jump Game II
This one was a challenging greedy problem. I couldn't get it-- I was
too preoccupied with the first Jump Game and couldn't think of a
unique solution for the second one.

The inspiration here is the same; it's greedy... then you need to
realize that you're working in bands. You want the maximum potential
reach in every band, and every time you reach a new band, you increment
jumps!

I had a bug because I did <= instead of <; you're supposed to end the
iteration when you reach the last index, you don't want to continue.