0134 Gas Station
Greedy problem! I guess the trick to problems like this is to identify
that it is a greedy problem... then figure out the trick.

Realize that if you have more gas than costs, if you can hit a "positive
streak" til' the end, the saved up positive numbers can complete the
circuit through all the negative. Honestly, that is absurdly not
intuitive... Then just save the first number of the positive streak.
So my initial intuition of solving the problem was right? Kind of
insane...