0763 Partition labels
A greedy string problem!
This was honestly super tricky, but I think the intuition is finding out
that it's greedy... then realizing basically
1. If you hit a letter, everything from that letter til its last
occurrence has to be together
2. For all the letters inbetween, they become "part" of that partition.
Meaning if for any letters within, they can set a new "end" if there's
a last occurrence farther in.
And... that's it.
Then you just count the size of each partition, and reset when you reach
a new partition (aka end of a partition is reached)