0678 Valid Parenthesis String
Greedy problem! I have no idea how you'd find this out... I guess
the intuition is that if left is ever greater than right (left comes
before right), it's invalid.

Then you realize * can change the min/max amount of lefts.
And that 0 lefts == it's valid
And that just ensuring that left max doesn't go negative (more rights
than lefts) is enough.
Since * can add to left max, or subtract from left min (if it becomes
a right.

This is honestly a super tricky greedy problem, it's kind of absurd.
And the backtracking or DP solutions are way more inefficient...