0200 Number of Islands and 0351 Android Unlock Patterns
0200 Number of Islands
- A graph flood fill problem
- Not much else to say, punny.. islands and floods

0351 Android Unlock Patterns
- Backtracking permutations problem
- Tricky part was figuring out how to do the skips
If thinking about it, it's logical, right? You have a current number
going to a future number, and you need to see if that number in the
middle has been visited or not. And you also need to know because
this only applies to starting and ending at certain numbers.
- Symmetry! Confused me a bit because symmetry only reduces the # of
calculations, not the complexity of the problem. You have to do all
skips either way, because even though you start at 1/2/5, you still need
to process all 9 and do whatever necessary skips for them.