0994 Rotting Oranges
A tricky multi-source BFS problem! I think for multi-source BFS
problems, you typically want to do all the work inside of the
directions loop, not outside of it, otherwise you end up double
adding and stuff like that. Also for this one, break if the new
queue is empty because that means no work is to be done, meaning no
need to increment.