0207 Course Schedule
A difficult graph problem!
Realizing that pairs are edges and courses are nodes.
THen thinking about how to properly detect cycles in an effective way.

Create an adjacency list (courses -> [prereqs])
DFS on courses, remove prereqs from an element if they can be reached
Set prereq to [] so future calls will just default to true

This graph + DFS problem relying on the intuition of detecting cycles
is actually pretty tricky-- you have to draw it out to understand.