0684 Redundant Connection
Fun graph problem! Once the problem is realized to be a cycle detection
problem, it gets a bit easier. Then the realization is that you can use
union find to detect cycles in graphs (if two components are already
connected) and you find an edge that connects them again, that's a
redundant connection!