feat(0139-word-break-rb-and-md): implemented 0139 word break
0139 Word Break is a very interesting DP problem. I did it both the
memoized and bottom-up DP ways.

The intuition for memoized is that with decision tree branches, repeat
work coudl be cached (true or false). Once you got the idea that each
word option results in a potential decision tree, it's not bad.

The intuition for bottom-up DP is a bit more tricky but makes sense. You
start from the back, and at each letter, you see if it can reach the
end. Thena ll you need to do when reaching a previous letter, is go
letter index + word.length and see if that reaches the end or reaches a
letter that reaches the end. It's essentially the same...

Problems like this require reading the problem very clearly and
honestly, whiteboarding would help a ton but too bad we don't get that
and just have to visualize.