feat(0191-number-of-1-bits): implement 0191 number of 1 bits
Fun bitwise operator problems with multiple ways to do it. I like the method of shifting by 1 bit to the right then checking if the last digit is a 1 or not. The "best" way relies on a trick by and'ing with n - 1 to skip all the zeroes, clever but I don't think I could come up with that quickly.