feat(0072 edit distance): implemented edit distance
Extremely hard 2D DP problem, but I think it's a classic. I think
a point of inspiration I'm lacking is honestly converting the base
cases and recurrence relations (subproblems) into a 2D graph. It's also
difficult for me to abstract away the string as a whole rather than
as individual chars.

E.g, when looking at base cases and setting how many operations...
[0][3] where out of bounds is 3 because for an empty string to be
the string "abc", it's 3 deletes

I have no idea why this problem was downgraded from a hard to a medium,
because it requires a huge amount of proper bottom-up DP intuition.