feat(312): implement 312 burst balloons
This was absurdly hard. 2D DP with intuition of bursting the balloon
last, and the memoizaiton solution isn't efficient enough so you have
to figure out the complicated tabulation solution.
Take [3, 1, 5, 8]
If you're bursting 3 last, the boundaries will be [1, 1]
For the [left, idx - 1] boundary, it's l > r so it returns 0 (empty)
For the [idx + 1, right] boundary, you go through all the choices
picking the max coins. If you choose 1 to pop first, notice how the
boundaries are 3 and 1 in this case. You're choosing every balloon to
pop LAST instead of first.