feat(0895): maximum Frequency Stack
Interesting stack problem, combining usage of hashmaps and stacks.

The intuition to solving this is trying to get it as efficient as
possible, even in cases where every element is the same. Essentially
going for O(1) pops and pushes. Since order matters, we can group
the elements by count. As an example:
1: [1, 2, 3, 4]
2: [1, 2, 3]
3: [1, 2]
4: [1]
Then we can just pop from the tracked max_freq everytime. We can add
to the proper count based on the character's count.

My mistake was not shifting perspectives-- I came up with a similar
solution but didn't utilize lists for proper ordering and came up
with complicated mechanisms that weren't as efficient.
I had:
Char: IDs of when they showed up
Count: Set of chars
Then I did max_freq the same way.
So I would go into count, get all of the chars for that count.
But since I didn't keep order, I'd go through each char comparing
the last index in which they showed up, popping it from that stack.
So whenever it was say, a lot of repeated chars in the low ranges,
I'd still have to do a practical O(N) search. I should have immediately
tried to think about another approach when my approach was still linear
in the worst case/average cases.